refactor: extract estimateAndTag to eliminate route duplication
- Move shared estimate+tag+patch workflow into estimateAndTag() in tagging.ts
- All three routes now call estimateAndTag instead of repeating the 8-line pattern
- Remove unused imports from routes
