+++ src/routes/estimate.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from "fastify"
-import { getRecipe, patchRecipe } from "../services/mealie-client.js"
-import { computeIngredientHash, estimateRecipe, buildNutritionPatch } from "../services/estimator.js"
-import { resolveAndMergeTags } from "../services/tagging.js"
+import { getRecipe } from "../services/mealie-client.js"
+import { computeIngredientHash } from "../services/estimator.js"
+import { estimateAndTag } from "../services/tagging.js"
 import { logger } from "../utils/logger.js"
 
 async function processEstimate(slug: string): Promise<void> {
@@ -10,17 +10,9 @@ async function processEstimate(slug: string): Promise<void> {
 
     const recipe = await getRecipe(slug)
     const hash = computeIngredientHash(recipe)
-    const result = await estimateRecipe(recipe)
-    const nutritionPatch = buildNutritionPatch(result, hash, recipe.recipeYield)
+    const { calories, tagSlugs } = await estimateAndTag(recipe, hash)
 
-    const { tags, tagSlugs } = await resolveAndMergeTags(recipe, result.perServingNutrients)
-    await patchRecipe(slug, {
-      ...nutritionPatch,
-      tags,
-      extras: { ...nutritionPatch.extras, calorie_estimator_tags: JSON.stringify(tagSlugs) },
-    })
-
-    logger.info({ slug, calories: result.perServingNutrients.kcalPer100g, tags: tagSlugs }, "On-demand estimation complete")
+    logger.info({ slug, calories, tags: tagSlugs }, "On-demand estimation complete")
   } catch (err) {
     logger.error({ slug, err }, "Estimate background processing failed")
   }
